% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_go_enrichment.r
\name{compare_go_enrichment}
\alias{compare_go_enrichment}
\title{Function to compare GO term enrichment in multiple lists of A. thaliana genes}
\usage{
compare_go_enrichment(
  agi_lists,
  variants = TRUE,
  relationships = "all",
  aspects = "all",
  evidence = "any",
  method = "binom",
  p.adjust_method = "BH",
  cutoff = 0.05,
  test_upstream = TRUE
)
}
\arguments{
\item{agi_lists}{A list of named vectors, each containing AGI codes of interest, for instance differentially expressed genes in different conditions.}

\item{variants}{A logical indicating whether or not isoform variants are included as part of their locus. Default: TRUE}

\item{relationships}{A vector of GO relationships to consider, see http://www.ontobee.org/ontology/RO .  Default: all}

\item{evidence}{A vector of GO evidence codes to consider, see http://geneontology.org/docs/guide-go-evidence-codes/ . Default: any}

\item{method}{The method to use for statistical testing; binom, fisher or chisq. Default: binom}

\item{p.adjust_method}{The method for p-value adjustment, see p.adjust. Default: BH}

\item{cutoff}{If given, only results with p-value less than or equal to this parameter will be reported. Default: 0.05}

\item{test_upstream}{A logical to determine if GO terms upstream of those given in ATH_GOSLIM will also be tested for significance. Default: TRUE}

\item{aspect}{A vector of GO aspects to consider. F: molecular function; C: cellular component; P: biological process. Default: all}
}
\value{
A labelled vector of adjusted p-values.
}
\description{
Function to compare GO term enrichment in multiple lists of A. thaliana genes
}
\details{
GO terms are found for all genes in the agi_list. For each term, associated genes are counted in the agi_list and in the rest of A. thaliana. A gene is considered associated if either it has a direct association according to ath_goslim or if it is associated with any of the offspring terms. A statistical test is then carried out to determine the likelihood of the term being over- or under-represented. P-values are then adjusted for multiple testing.
}
\examples{
None
}
\author{
Chris Field \href{mailto:fieldc@ethz.ch}{fieldc@ethz.ch}
}
