% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_go_comparison.r
\name{plot_go_comparison}
\alias{plot_go_comparison}
\title{Function to plot a heatmap of a comparison of GO term enrichment between conditions}
\usage{
plot_go_comparison(p_value_table, ...)
}
\arguments{
\item{p_value_table}{A data frame of p-values created by compare_go_enrichment()}

\item{...}{Other arguments passed to heatmap()}
}
\value{
None.
}
\description{
Function to plot a heatmap of a comparison of GO term enrichment between conditions
}
\details{

}
\examples{
None
}
\author{
Chris Field \href{mailto:fieldc@ethz.ch}{fieldc@ethz.ch}
}
